@@ -28,9 +28,10 @@ from cvxpy.problems.problem import Problem
 from cvxpy.transforms import linearize, partial_optimize, suppfunc
 from cvxpy.reductions import *
 from cvxpy.reductions.solvers.defines import installed_solvers
-from cvxpy.settings import (CBC, CPLEX, CVXOPT, DIFFCP, ECOS, ECOS_BB, GLPK,
+from cvxpy.settings import (CBC, CLARABEL, COPT, CPLEX, CVXOPT, DIFFCP, ECOS, ECOS_BB, GLPK,
                             GLPK_MI, GUROBI, INFEASIBLE, INFEASIBLE_INACCURATE,
-                            MOSEK, NAG, OPTIMAL, OPTIMAL_INACCURATE, OSQP,
-                            ROBUST_KKTSOLVER, SCIP, SCIPY, SCS, SOLVER_ERROR,
-                            UNBOUNDED, UNBOUNDED_INACCURATE, USER_LIMIT,
-                            XPRESS, get_num_threads, set_num_threads,)
+                            MOSEK, NAG, PDLP, OPTIMAL, OPTIMAL_INACCURATE, OSQP,
+                            ROBUST_KKTSOLVER, GLOP, SCIP, SCIPY, SCS, SDPA,
+                            SOLVER_ERROR, UNBOUNDED, UNBOUNDED_INACCURATE,
+                            USER_LIMIT, XPRESS, get_num_threads,
+                            set_num_threads,)
