@@ -88,7 +88,7 @@ class AxisAtom(Atom):
                     if d is None:
                         return [None]
                     else:
-                        d = d.flatten()
+                        d = np.array(d).flatten()
                     row = np.linspace(i*n, i*n+m-1, m)  # [i*n, i*n+1, ..., i*n+m-1]
                     col = np.ones((m))*i
                     D = D + sp.csc_matrix((d, (row, col)),
