@@ -17,6 +17,7 @@ from typing import Tuple
 
 import numpy as np
 
+import cvxpy.settings as s
 from cvxpy.atoms.atom import Atom
 
 
@@ -33,7 +34,8 @@ class length(Atom):
     def numeric(self, values) -> int:
         """Returns the length of x.
         """
-        return np.max(np.nonzero(values[0])) + 1
+        outside_tol = np.abs(values[0]) > s.ATOM_EVAL_TOL
+        return np.max(np.nonzero(outside_tol)) + 1
 
     def shape_from_args(self) -> Tuple[int, ...]:
         """Returns the (row, col) shape of the expression.
