@@ -14,7 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from cvxpy.constraints.exponential import ExpCone
+from cvxpy.constraints.exponential import ExpCone, OpRelConeQuad, RelEntrQuad
+from cvxpy.constraints.finite_set import FiniteSet
 from cvxpy.constraints.nonpos import Inequality, NonNeg, NonPos
 from cvxpy.constraints.power import PowCone3D, PowConeND
 from cvxpy.constraints.psd import PSD
