@@ -28,6 +28,7 @@ from cvxpy import settings as s
 from cvxpy.atoms.atom import Atom
 from cvxpy.constraints import Equality, Inequality, NonNeg, NonPos, Zero
 from cvxpy.constraints.constraint import Constraint
+from cvxpy.error import DPPError
 from cvxpy.expressions import cvxtypes
 from cvxpy.expressions.variable import Variable
 from cvxpy.interface.matrix_utilities import scalar_value
@@ -48,6 +49,7 @@ from cvxpy.reductions.solvers.solver import Solver
 from cvxpy.reductions.solvers.solving_chain import (SolvingChain,
                                                     construct_solving_chain,)
 from cvxpy.settings import SOLVERS
+from cvxpy.utilities import debug_tools
 from cvxpy.utilities.deterministic import unique_list
 
 SolveResult = namedtuple(
@@ -101,8 +103,8 @@ class Cache:
         self.param_prog = None
         self.inverse_data = None
 
-    def make_key(self, solver, gp):
-        return (solver, gp)
+    def make_key(self, solver, gp, ignore_dpp):
+        return (solver, gp, ignore_dpp)
 
     def gp(self):
         return self.key is not None and self.key[1]
@@ -147,7 +149,17 @@ class Problem(u.Canonical):
             raise error.DCPError("Problem objective must be Minimize or Maximize.")
         # Constraints and objective are immutable.
         self._objective = objective
+        # Raise warning if objective has too many subexpressions.
+        if debug_tools.node_count(self._objective) >= debug_tools.MAX_NODES:
+            warnings.warn("Objective contains too many subexpressions. "
+                          "Consider vectorizing your CVXPY code to speed up compilation.")
         self._constraints = [_validate_constraint(c) for c in constraints]
+        # Raise warning if constraint has too many subexpressions.
+        for i, constraint in enumerate(self._constraints):
+            if debug_tools.node_count(constraint) >= debug_tools.MAX_NODES:
+                warnings.warn(f"Constraint #{i} contains too many subexpressions. "
+                              "Consider vectorizing your CVXPY code to speed up compilation.")
+
         self._value = None
         self._status: Optional[str] = None
         self._solution = None
@@ -425,6 +437,9 @@ class Problem(u.Canonical):
             When True, a DPPError will be thrown when trying to solve a non-DPP
             problem (instead of just a warning). Only relevant for problems
             involving Parameters. Defaults to False.
+        ignore_dpp : bool, optional
+            When True, DPP problems will be treated as non-DPP,
+            which may speed up compilation. Defaults to False.
         method : function, optional
             A custom solve method to use.
         kwargs : keywords, optional
@@ -463,6 +478,8 @@ class Problem(u.Canonical):
             Raised if the problem is not DCP and `gp` is False.
         cvxpy.error.DGPError
             Raised if the problem is not DGP and `gp` is True.
+        cvxpy.error.DPPError
+            Raised if DPP settings are invalid.
         cvxpy.error.SolverError
             Raised if no suitable solver exists among the installed solvers,
             or if an unanticipated error is encountered.
@@ -489,7 +506,12 @@ class Problem(u.Canonical):
         cls.REGISTERED_SOLVE_METHODS[name] = func
 
     def get_problem_data(
-        self, solver, gp: bool = False, enforce_dpp: bool = False, verbose: bool = False
+        self, solver,
+        gp: bool = False,
+        enforce_dpp: bool = False,
+        ignore_dpp: bool = False,
+        verbose: bool = False,
+        solver_opts: Optional[dict] = None
     ):
         """Returns the problem data used in the call to the solver.
 
@@ -564,8 +586,15 @@ class Problem(u.Canonical):
         enforce_dpp : bool, optional
             When True, a DPPError will be thrown when trying to parse a non-DPP
             problem (instead of just a warning). Defaults to False.
+        ignore_dpp : bool, optional
+            When True, DPP problems will be treated as non-DPP,
+            which may speed up compilation. Defaults to False.
         verbose : bool, optional
             If True, print verbose output related to problem compilation.
+        solver_opts : dict, optional
+            A dict of options that will be passed to the specific solver.
+            In general, these options will override any default settings
+            imposed by cvxpy.
 
         Returns
         -------
@@ -575,13 +604,27 @@ class Problem(u.Canonical):
             The solving chain that created the data.
         list
             The inverse data generated by the chain.
+
+        Raises
+        ------
+        cvxpy.error.DPPError
+            Raised if DPP settings are invalid.
         """
+        # Invalid DPP setting.
+        # Must be checked here to avoid cache issues.
+        if enforce_dpp and ignore_dpp:
+            raise DPPError("Cannot set enforce_dpp = True and ignore_dpp = True.")
+
         start = time.time()
-        key = self._cache.make_key(solver, gp)
+        # Cache includes ignore_dpp because it alters compilation.
+        key = self._cache.make_key(solver, gp, ignore_dpp)
         if key != self._cache.key:
             self._cache.invalidate()
             solving_chain = self._construct_chain(
-                solver=solver, gp=gp, enforce_dpp=enforce_dpp)
+                solver=solver, gp=gp,
+                enforce_dpp=enforce_dpp,
+                ignore_dpp=ignore_dpp,
+                solver_opts=solver_opts)
             self._cache.key = key
             self._cache.solving_chain = solving_chain
             self._solver_cache = {}
@@ -724,12 +767,16 @@ class Problem(u.Canonical):
                     for discussion on this topic.
 
                     Quick fix 1: if you install the python package CVXOPT (pip install cvxopt),
-                    then CVXPY can use the open-source mixed-integer solver `GLPK`.
+                    then CVXPY can use the open-source mixed-integer linear programming
+                    solver `GLPK`. If your problem is nonlinear then you can install SCIP
+                    (pip install pyscipopt).
 
                     Quick fix 2: you can explicitly specify solver='ECOS_BB'. This may result
                     in incorrect solutions and is not recommended.
                 """
                 raise error.SolverError(msg)
+            # TODO: provide a useful error message when the problem is nonlinear but
+            #  the only installed mixed-integer solvers are MILP solvers (e.g., GLPK_MI).
             candidates['qp_solvers'] = [
                 s for s in candidates['qp_solvers']
                 if slv_def.SOLVER_MAP_QP[s].MIP_CAPABLE]
@@ -768,7 +815,7 @@ class Problem(u.Canonical):
         """
         if custom_solver.name() in SOLVERS:
             message = "Custom solvers must have a different name than the officially supported ones"
-            raise(error.SolverError(message))
+            raise error.SolverError(message)
 
         candidates = {'qp_solvers': [], 'conic_solvers': []}
         if not self.is_mixed_integer() or custom_solver.MIP_CAPABLE:
@@ -781,7 +828,9 @@ class Problem(u.Canonical):
         return candidates
 
     def _construct_chain(
-        self, solver: Optional[str] = None, gp: bool = False, enforce_dpp: bool = False
+        self, solver: Optional[str] = None, gp: bool = False,
+        enforce_dpp: bool = False, ignore_dpp: bool = False,
+        solver_opts: Optional[dict] = None
     ) -> SolvingChain:
         """
         Construct the chains required to reformulate and solve the problem.
@@ -801,6 +850,11 @@ class Problem(u.Canonical):
             instead of as a Disciplined Convex Program.
         enforce_dpp : bool, optional
             Whether to error on DPP violations.
+        ignore_dpp : bool, optional
+            When True, DPP problems will be treated as non-DPP,
+            which may speed up compilation. Defaults to False.
+        solver_opts: dict, optional
+            Additional arguments to pass to the solver.
 
         Returns
         -------
@@ -809,7 +863,9 @@ class Problem(u.Canonical):
         candidate_solvers = self._find_candidate_solvers(solver=solver, gp=gp)
         self._sort_candidate_solvers(candidate_solvers)
         return construct_solving_chain(self, candidate_solvers, gp=gp,
-                                       enforce_dpp=enforce_dpp)
+                                       enforce_dpp=enforce_dpp,
+                                       ignore_dpp=ignore_dpp,
+                                       solver_opts=solver_opts)
 
     @staticmethod
     def _sort_candidate_solvers(solvers) -> None:
@@ -847,6 +903,7 @@ class Problem(u.Canonical):
                qcp: bool = False,
                requires_grad: bool = False,
                enforce_dpp: bool = False,
+               ignore_dpp: bool = False,
                **kwargs):
         """Solves a DCP compliant optimization problem.
 
@@ -874,6 +931,9 @@ class Problem(u.Canonical):
         enforce_dpp : bool, optional
             When True, a DPPError will be thrown when trying to solve a non-DPP
             problem (instead of just a warning). Defaults to False.
+        ignore_dpp : bool, optional
+            When True, DPP problems will be treated as non-DPP,
+            which may speed up compilation. Defaults to False.
         kwargs : dict, optional
             A dict of options that will be passed to the specific solver.
             In general, these options will override any default settings
@@ -966,7 +1026,7 @@ class Problem(u.Canonical):
                 return self.value
 
         data, solving_chain, inverse_data = self.get_problem_data(
-            solver, gp, enforce_dpp, verbose)
+            solver, gp, enforce_dpp, ignore_dpp, verbose, kwargs)
 
         if verbose:
             print(_NUM_SOLVER_STR)
