@@ -26,7 +26,13 @@ def cumsum_canon(expr, args):
     # X = Y[1:,:] - Y[:-1, :]
     Y = Variable(expr.shape)
     if axis == 0:
-        constr = [X[1:] == Y[1:] - Y[:-1], Y[0] == X[0]]
+        if (expr.shape[0] == 1):
+            return X, []
+        else:
+            constr = [X[1:] == Y[1:] - Y[:-1], Y[0] == X[0]]
     else:
-        constr = [X[:, 1:] == Y[:, 1:] - Y[:, :-1], Y[:, 0] == X[:, 0]]
-    return (Y, constr)
+        if (expr.shape[1] == 1):
+            return X, []
+        else:
+            constr = [X[:, 1:] == Y[:, 1:] - Y[:, :-1], Y[:, 0] == X[:, 0]]
+    return Y, constr
