@@ -24,11 +24,10 @@ from cvxpy.reductions.qp2quad_form.atom_canonicalizers.quad_form_canon import *
 from cvxpy.reductions.qp2quad_form.atom_canonicalizers.quad_over_lin_canon import *
 from cvxpy.transforms.indicator import indicator
 
-CANON_METHODS = {}
-
 # TODO: remove pwl canonicalize methods, use EliminatePwl reduction instead
 
-# reuse cone canonicalization methods
+# Conic canonicalization methods.
+CANON_METHODS = {}
 CANON_METHODS[abs] = CONE_METHODS[abs]
 CANON_METHODS[cumsum] = CONE_METHODS[cumsum]
 CANON_METHODS[maximum] = CONE_METHODS[maximum]
@@ -41,8 +40,12 @@ CANON_METHODS[norm_inf] = CONE_METHODS[norm_inf]
 CANON_METHODS[indicator] = CONE_METHODS[indicator]
 CANON_METHODS[special_index] = CONE_METHODS[special_index]
 
-# canonicalizations that are different for QPs
-CANON_METHODS[quad_over_lin] = quad_over_lin_canon
-CANON_METHODS[power] = power_canon
-CANON_METHODS[huber] = huber_canon
-CANON_METHODS[QuadForm] = quad_form_canon
+# Canonicalizations that return a quadratic objective.
+# Saved here for reference in other files.
+QUAD_CANON_METHODS = {}
+QUAD_CANON_METHODS[quad_over_lin] = quad_over_lin_canon
+QUAD_CANON_METHODS[power] = power_canon
+QUAD_CANON_METHODS[huber] = huber_canon
+QUAD_CANON_METHODS[QuadForm] = quad_form_canon
+
+CANON_METHODS.update(QUAD_CANON_METHODS)
