@@ -28,6 +28,7 @@ from cvxpy.reductions.solution import Solution, failure_solution
 from cvxpy.reductions.solvers import utilities
 from cvxpy.reductions.solvers.conic_solvers.conic_solver import (
     ConicSolver, dims_to_solver_dict,)
+from cvxpy.utilities.versioning import Version
 
 log = logging.getLogger(__name__)
 
@@ -80,6 +81,7 @@ class SCIP(ConicSolver):
 
     MIP_CAPABLE = True
     SUPPORTED_CONSTRAINTS = ConicSolver.SUPPORTED_CONSTRAINTS + [SOC]
+    MI_SUPPORTED_CONSTRAINTS = SUPPORTED_CONSTRAINTS
 
     def name(self) -> str:
         """The name of the solver."""
@@ -87,8 +89,12 @@ class SCIP(ConicSolver):
 
     def import_solver(self) -> None:
         """Imports the solver."""
-        from pyscipopt import scip
-        scip  # For flake8
+        import pyscipopt
+        v = pyscipopt.__version__
+        if Version(v) >= Version('4.0.0'):
+            msg = 'PySCIPOpt (SCIP\'s Python wrapper) is installed and its ' \
+                  'version is %s. CVXPY only supports PySCIPOpt < 4.0.0.' % v
+            raise ModuleNotFoundError(msg)
 
     def apply(self, problem: ParamConeProg) -> Tuple[Dict, Dict]:
         """Returns a new problem and data for inverting the new solution."""
@@ -309,10 +315,14 @@ class SCIP(ConicSolver):
     ) -> Dict[str, Any]:
         """Solve and return a solution if one exists."""
 
-        solution = {}
         try:
             model.optimize()
+        except Exception as e:
+            log.warning("Error encountered when optimising %s: %s", model, e)
 
+        solution = {}
+
+        if max(model.getNSols(), model.getNCountedSols()) > 0:
             solution["value"] = model.getObjVal()
             sol = model.getBestSol()
             solution["primal"] = array([sol[v] for v in variables])
@@ -337,9 +347,6 @@ class SCIP(ConicSolver):
                 solution["y"] = -array(vals)
                 solution[s.EQ_DUAL] = solution["y"][0:dims[s.EQ_DIM]]
                 solution[s.INEQ_DUAL] = solution["y"][dims[s.EQ_DIM]:]
-
-        except Exception as e:
-            log.warning("Error encountered when optimising %s: %s", model, e)
 
         solution[s.SOLVE_TIME] = model.getSolvingTime()
         solution['status'] = STATUS_MAP[model.getStatus()]
