@@ -3,6 +3,7 @@ import scipy.sparse as sp
 
 import cvxpy.interface as intf
 import cvxpy.settings as s
+from cvxpy.error import SolverError
 from cvxpy.reductions.solution import Solution, failure_solution
 from cvxpy.reductions.solvers.qp_solvers.qp_solver import QpSolver
 
@@ -79,7 +80,7 @@ class OSQP(QpSolver):
                 P_triu = sp.triu(P).tocsc()
                 new_args['Px'] = P_triu.data
                 factorizing = True
-            if A.data.shape != old_data['Ax'].shape or any(
+            if A.data.shape != old_data['Ax'].data.shape or any(
                     A.data != old_data['Ax'].data):
                 new_args['Ax'] = A.data
                 factorizing = True
@@ -97,7 +98,10 @@ class OSQP(QpSolver):
             # Initialize and solve problem
             solver_opts['polish'] = solver_opts.get('polish', True)
             solver = osqp.OSQP()
-            solver.setup(P, q, A, lA, uA, verbose=verbose, **solver_opts)
+            try:
+                solver.setup(P, q, A, lA, uA, verbose=verbose, **solver_opts)
+            except ValueError as e:
+                raise SolverError(e)
 
         results = solver.solve()
 
