@@ -21,6 +21,7 @@ import pytest
 import scipy
 import scipy.sparse as sp
 import scipy.stats
+from numpy import linalg as LA
 
 import cvxpy as cp
 import cvxpy.settings as s
@@ -110,6 +111,15 @@ class TestAtoms(BaseTest):
             "The input must be a single CVXPY Expression, not a list. "
             "Combine Expressions using atoms such as bmat, hstack, and vstack."))
 
+    def test_norm_exceptions(self) -> None:
+        """Test that norm exceptions are raised as expected.
+        """
+        x = cp.Variable(2)
+        with self.assertRaises(Exception) as cm:
+            cp.norm(x, 'nuc')
+        self.assertTrue(str(cm.exception) in (
+            "Unsupported norm option nuc for non-matrix."))
+
     def test_quad_form(self) -> None:
         """Test quad_form atom.
         """
@@ -159,6 +169,20 @@ class TestAtoms(BaseTest):
                 self.assertEqual(copy.get_data(), atom.get_data())
 
         assert cp.power(-1, 2).value == 1
+
+    # Test the xexp class
+    def test_xexp(self) -> None:
+        # Test for positive x
+        x = Variable(pos=True)
+        atom = cp.xexp(x)
+        self.assertEqual(atom.curvature, s.CONVEX)
+        self.assertEqual(atom.sign, s.NONNEG)
+
+        # Test for negative x
+        x = Variable(neg=True)
+        atom = cp.xexp(x)
+        self.assertNotEqual(atom.curvature, s.CONCAVE)
+        self.assertEqual(atom.sign, s.NONPOS)
 
     # Test the geo_mean class.
     def test_geo_mean(self) -> None:
@@ -336,8 +360,10 @@ class TestAtoms(BaseTest):
         # Invalid axis.
         with self.assertRaises(Exception) as cm:
             cp.max(self.x, axis=4)
-        self.assertEqual(str(cm.exception),
-                         "Invalid argument for axis.")
+        self.assertEqual(str(cm.exception), "Invalid argument for axis.")
+        with self.assertRaises(ValueError) as cm:
+            cp.max(self.x, self.x)  # a common erroneous use-case
+        self.assertEqual(str(cm.exception), cp.max.__EXPR_AXIS_ERROR__)
 
     def test_min(self) -> None:
         """Test min.
@@ -357,8 +383,10 @@ class TestAtoms(BaseTest):
         # Invalid axis.
         with self.assertRaises(Exception) as cm:
             cp.min(self.x, axis=4)
-        self.assertEqual(str(cm.exception),
-                         "Invalid argument for axis.")
+        self.assertEqual(str(cm.exception), "Invalid argument for axis.")
+        with self.assertRaises(ValueError) as cm:
+            cp.min(self.x, self.x)  # a common erroneous use-case
+        self.assertEqual(str(cm.exception), cp.min.__EXPR_AXIS_ERROR__)
 
     # Test sign logic for maximum.
     def test_maximum_sign(self) -> None:
@@ -628,6 +656,18 @@ class TestAtoms(BaseTest):
         self.assertEqual(str(cm.exception),
                          "Argument to trace must be a square matrix.")
 
+    def test_trace_sign_psd(self) -> None:
+        """Test sign of trace for psd/nsd inputs.
+        """
+        X_psd = cp.Variable((2, 2), PSD=True)
+        X_nsd = cp.Variable((2, 2), NSD=True)
+
+        psd_trace = cp.trace(X_psd)
+        nsd_trace = cp.trace(X_nsd)
+
+        assert psd_trace.is_nonneg()
+        assert nsd_trace.is_nonpos()
+
     def test_log1p(self) -> None:
         """Test the log1p atom.
         """
@@ -818,7 +858,7 @@ class TestAtoms(BaseTest):
         self.assertEqual(str(cm.exception),
                          "The arguments to conv must resolve to vectors.")
 
-    def test_kron(self) -> None:
+    def test_kron_expr(self) -> None:
         """Test the kron atom.
         """
         a = np.ones((3, 2))
@@ -830,9 +870,9 @@ class TestAtoms(BaseTest):
         expr = cp.kron(a, b)
         assert expr.is_nonpos()
         with self.assertRaises(Exception) as cm:
-            cp.kron(self.x, -1)
+            cp.kron(self.x, self.x)
         self.assertEqual(str(cm.exception),
-                         "The first argument to kron must be constant.")
+                         "At least one argument to kron must be constant.")
 
     def test_partial_optimize_dcp(self) -> None:
         """Test DCP properties of partial optimize.
@@ -1100,6 +1140,19 @@ class TestAtoms(BaseTest):
         self.assertEqual(cp.diff(A, axis=1).shape,
                          np.diff(B, axis=1).shape)
 
+        # Issue #1834
+
+        x1 = np.array([[1, 2, 3, 4, 5]])
+        x2 = cp.Variable((1, 5), value=x1)
+
+        expr = cp.diff(x1, axis=1)
+        self.assertItemsAlmostEqual(expr.value, np.diff(x1, axis=1))
+        expr = cp.diff(x2, axis=1)
+        self.assertItemsAlmostEqual(expr.value, np.diff(x1, axis=1))
+
+        with pytest.raises(ValueError, match="< k elements"):
+            cp.diff(x1, axis=0).value
+
     def test_log_normcdf(self) -> None:
         self.assertEqual(cp.log_normcdf(self.x).sign, s.NONPOS)
         self.assertEqual(cp.log_normcdf(self.x).curvature, s.CONCAVE)
@@ -1287,3 +1340,350 @@ class TestAtoms(BaseTest):
             cp.partial_transpose(X, dims=[2, 4], axis=0)
         self.assertEqual(str(cm.exception),
                          "Dimension of system doesn't correspond to dimension of subsystems.")
+
+    def test_log_sum_exp(self) -> None:
+        """Test log_sum_exp sign.
+        """
+        # Test for non-negative x
+        x = Variable(nonneg=True)
+        atom = cp.log_sum_exp(x)
+        self.assertEqual(atom.curvature, s.CONVEX)
+        self.assertEqual(atom.sign, s.NONNEG)
+
+        # Test for non-positive x
+        x = Variable(nonpos=True)
+        atom = cp.log_sum_exp(x)
+        self.assertEqual(atom.curvature, s.CONVEX)
+        self.assertEqual(atom.sign, s.UNKNOWN)
+
+    def test_flatten(self) -> None:
+        """Test flatten and vec."""
+        # Constant argument.
+        A = np.arange(10)
+        reshaped = np.reshape(A, (2, 5), order='F')
+        expr = cp.vec(reshaped, order='F')
+        self.assertItemsAlmostEqual(expr.value, A)
+        expr = cp.Constant(reshaped).flatten(order='F')
+        self.assertItemsAlmostEqual(expr.value, A)
+
+        reshaped = np.reshape(A, (2, 5), order='C')
+        expr = cp.vec(reshaped, order='C')
+        self.assertItemsAlmostEqual(expr.value, A)
+        expr = cp.Constant(reshaped).flatten(order='C')
+        self.assertItemsAlmostEqual(expr.value, A)
+
+        reshaped = np.reshape(A, (2, 5), order='F')
+        expr = cp.vec(reshaped, order='F')
+        self.assertItemsAlmostEqual(expr.value, A)
+        expr = cp.Constant(reshaped).flatten()
+        self.assertItemsAlmostEqual(expr.value, A)
+
+        # Variable argument.
+        x = Variable((2, 5))
+        reshaped = np.reshape(A, (2, 5), order='F')
+        expr = cp.vec(x, order='F')
+        cp.Problem(cp.Minimize(0), [expr == A]).solve()
+        self.assertItemsAlmostEqual(x.value, reshaped)
+        expr = cp.Constant(A).flatten(order='F')
+        cp.Problem(cp.Minimize(0), [expr == A]).solve()
+        self.assertItemsAlmostEqual(x.value, reshaped)
+
+        reshaped = np.reshape(A, (2, 5), order='C')
+        expr = cp.vec(x, order='C')
+        cp.Problem(cp.Minimize(0), [expr == A]).solve()
+        self.assertItemsAlmostEqual(x.value, reshaped)
+        expr = cp.Constant(A).flatten(order='C')
+        cp.Problem(cp.Minimize(0), [expr == A]).solve()
+        self.assertItemsAlmostEqual(x.value, reshaped)
+
+        reshaped = np.reshape(A, (2, 5), order='F')
+        expr = cp.vec(x)
+        cp.Problem(cp.Minimize(0), [expr == A]).solve()
+        self.assertItemsAlmostEqual(x.value, reshaped)
+        expr = cp.Constant(A).flatten()
+        cp.Problem(cp.Minimize(0), [expr == A]).solve()
+        self.assertItemsAlmostEqual(x.value, reshaped)
+
+    def test_tr_inv(self) -> None:
+        """Test tr_inv atom. """
+        T = 5
+        # Solves the following SDP problem:
+        #           minimize    trace(inv(X))
+        #               s.t.    X is PSD
+        #                       trace(X)==1
+
+        # Create a symmetric matrix variable.
+        X = cp.Variable((T, T), symmetric=True)
+
+        # Define and solve the CVXPY problem.
+        # X should be a PSD
+        constraints = [X >> 0]
+        constraints += [
+            cp.trace(X) == 1
+        ]
+        prob = cp.Problem(cp.Minimize(cp.tr_inv(X)), constraints)
+        prob.solve(verbose=True)
+        # Check result. The best value is T^2.
+        self.assertAlmostEqual(prob.value, T**2)
+        X_actual = X.value
+        X_expect = np.eye(T) / T
+        self.assertItemsAlmostEqual(X_actual, X_expect, places=4)
+        # Second SDP problem, given a row full-rank matrix M:
+        #           minimize    trace(inv(M * X * M.T))
+        #               s.t.    X is PSD
+        #                       -1 <= X[i][j] <= 1 for all i,j
+        constraints = [X >> 0]
+        # n should not be greater than T,
+        # because the input should be positive definite.
+        n = 4
+        M = np.random.randn(n, T)
+        constraints += [
+            X >= -1,
+            X <= 1,
+        ]
+        prob = cp.Problem(cp.Minimize(cp.tr_inv(M @ X @ M.T)), constraints)
+        MM = M @ M.T
+        naiveRes = np.sum(LA.eigvalsh(MM) ** -1)
+        prob.solve(verbose=True)
+        # The optimized result should be smaller than the naive result,
+        # where X of the naive result is I.
+        self.assertTrue(prob.value < naiveRes)
+
+
+class TestDotsort(BaseTest):
+    """ Unit tests for the dotsort atom. """
+
+    def setUp(self) -> None:
+        self.x = cp.Variable(5)
+
+    def test_sum_k_largest_equivalence(self):
+        x_val = np.array([1, 3, 2, -5, 0])
+        w = np.array([1, 1, 1, 0])
+        expr = cp.dotsort(self.x, w)
+        assert expr.is_convex()
+        assert expr.is_incr(0)
+        prob = cp.Problem(cp.Minimize(expr), [self.x == x_val])
+        prob.solve()
+        self.assertAlmostEqual(prob.objective.value, np.sum(np.sort(x_val)[-3:]))
+
+    def test_sum_k_smallest_equivalence(self):
+        x_val = np.array([1, 3, 2, -5, 0])
+        w = np.array([-1, -1, -1, 0])
+        expr = -cp.dotsort(self.x, w)
+        assert expr.is_concave()
+        assert expr.is_decr(0)
+        prob = cp.Problem(cp.Maximize(expr), [self.x == x_val])
+        prob.solve()
+        self.assertAlmostEqual(prob.objective.value, np.sum(np.sort(x_val)[:3]))
+
+    def test_1D(self):
+        x_val = np.array([1, 3, 2, -5, 0])
+        w = np.array([-1, 5, 2, 0, 5])
+        expr = cp.dotsort(self.x, w)
+        assert expr.is_convex()
+        assert not expr.is_incr(0)
+        assert not expr.is_decr(0)
+
+        prob = cp.Problem(cp.Minimize(expr), [self.x == x_val])
+        prob.solve()
+        self.assertAlmostEqual(prob.objective.value, np.sort(x_val) @ np.sort(w))
+
+    def test_2D(self):
+        # X, W are matrix valued
+        x = cp.Variable((5, 5))
+        x_val = np.arange(25).reshape((5, 5))
+        w = np.arange(4).reshape((2, 2))
+        w_padded = np.zeros_like(x_val)
+        w_padded[:w.shape[0], :w.shape[1]] = w
+        expr = cp.dotsort(x, w)
+        assert expr.is_convex()
+        assert expr.is_incr(0)
+
+        prob = cp.Problem(cp.Minimize(expr), [x == x_val])
+        prob.solve()
+        self.assertAlmostEqual(prob.objective.value,
+                               np.sort(x_val.flatten()) @ np.sort(w_padded.flatten()))
+
+    def test_0D(self):
+        # scalar w
+        x_val = np.array([1, 3, 2, -5, 0])
+        w = 1
+        expr = cp.dotsort(self.x, w)
+        assert expr.is_convex()
+        assert expr.is_incr(0)
+        prob = cp.Problem(cp.Minimize(expr), [self.x == x_val])
+        prob.solve()
+        self.assertAlmostEqual(prob.objective.value, np.sum(np.sort(x_val)[-1:]))
+
+        # scalar x, w
+        x = cp.Variable()
+        x_val = np.array([1])
+        w = 1
+        expr = cp.dotsort(x, w)
+        assert expr.is_convex()
+        assert expr.is_incr(0)
+        prob = cp.Problem(cp.Minimize(expr), [x == x_val])
+        prob.solve()
+        self.assertAlmostEqual(prob.objective.value, np.sum(np.sort(x_val)[-1:]))
+
+    def test_constant(self):
+        x = np.arange(25)
+        x_val = np.arange(25).reshape((5, 5))
+        w = np.arange(4).reshape((2, 2))
+        w_padded = np.zeros_like(x_val)
+        w_padded[:w.shape[0], :w.shape[1]] = w
+        expr = cp.dotsort(x, w)
+        assert expr.is_convex()
+        assert expr.is_incr(0)
+
+        prob = cp.Problem(cp.Minimize(expr), [])
+        prob.solve()
+        self.assertAlmostEqual(prob.objective.value,
+                               np.sort(x_val.flatten()) @ np.sort(w_padded.flatten()))
+
+    def test_parameter(self):
+        x_val = np.array([1, 3, 2, -5, 0])
+
+        assert cp.dotsort(self.x, cp.Parameter(2, pos=True)).is_incr(0)
+        assert cp.dotsort(self.x, cp.Parameter(2, nonneg=True)).is_incr(0)
+        assert not cp.dotsort(self.x, cp.Parameter(2, neg=True)).is_incr(0)
+        assert cp.dotsort(self.x, cp.Parameter(2, neg=True)).is_decr(0)
+
+        w_p = cp.Parameter(2, value=[1, 0])
+        expr = cp.dotsort(self.x, w_p)
+        assert not expr.is_incr(0)
+        assert not expr.is_decr(0)
+
+        prob = cp.Problem(cp.Minimize(expr), [self.x == x_val])
+        prob.solve(enforce_dpp=True)
+        self.assertAlmostEqual(prob.objective.value,
+                               np.sort(x_val) @ np.sort(np.array([1, 0, 0, 0, 0])))
+        w_p.value = [-1, -1]
+        prob.solve(enforce_dpp=True)
+        self.assertAlmostEqual(prob.objective.value,
+                               np.sort(x_val) @ np.sort(np.array([-1, -1, 0, 0, 0])))
+
+        # Test parameter affine
+        w_p = cp.Parameter(2, value=[1, 0])
+        parameter_affine_expression = 2 * w_p
+        expr = cp.dotsort(self.x, parameter_affine_expression)
+        prob = cp.Problem(cp.Minimize(expr), [self.x == x_val])
+        prob.solve(enforce_dpp=True)
+        self.assertAlmostEqual(prob.objective.value,
+                               np.sort(x_val) @ np.sort(np.array([2, 0, 0, 0, 0])))
+        w_p.value = [-1, -1]
+        prob.solve(enforce_dpp=True)
+        self.assertAlmostEqual(prob.objective.value,
+                               np.sort(x_val) @ np.sort(np.array([-2, -2, 0, 0, 0])))
+
+        # Test constant + non-parameter affine
+        x_const = np.array([1, 2, 3])
+        p = cp.Parameter(value=2)
+        p_squared = p ** 2
+        expr = cp.dotsort(x_const, p_squared)
+        problem = cp.Problem(cp.Minimize(expr))
+        problem.solve(enforce_dpp=True)
+        self.assertAlmostEqual(expr.value, 2 ** 2 * 3)
+        p.value = -1
+        problem.solve(enforce_dpp=True)
+        self.assertAlmostEqual(expr.value, (-1) ** 2 * 3)
+
+        # Test non-parameter affine w/ with implicit casting to constant
+        with pytest.warns(UserWarning,
+                          match='You are solving a parameterized problem that is not DPP.'):
+            x_val = np.array([1, 2, 3, 4, 5])
+            p = cp.Parameter(value=2)
+            p_squared = p ** 2
+            expr = cp.dotsort(self.x, p_squared)
+            problem = cp.Problem(cp.Minimize(expr), [self.x == x_val])
+            problem.solve()
+            self.assertAlmostEqual(expr.value, 2 ** 2 * 5)
+            p.value = -1
+            problem.solve()
+            self.assertAlmostEqual(expr.value, (-1) ** 2 * 5)
+
+    def test_list(self):
+        r = np.array([2, 1, 0, -1, -1])
+        w = [1.2, 1.1]
+        expr = cp.dotsort(self.x, w)
+        prob = cp.Problem(cp.Maximize(r @ self.x), [0 <= self.x, expr <= 1, cp.sum(self.x) == 1])
+        prob.solve()
+        self.assertAlmostEqual(expr.value, 1)
+        self.assertAlmostEqual(self.x.value[:2] @ w, 1)
+
+    def test_composition(self):
+        r = np.array([2, 1, 0, -1, -1])
+        w = [0.7, 0.8]
+        expr = cp.dotsort(cp.exp(self.x), w)
+        prob = cp.Problem(cp.Maximize(r @ self.x), [0 <= self.x, expr <= 2, cp.sum(self.x) == 1])
+        prob.solve()
+        self.assertAlmostEqual(expr.value, 2)
+        self.assertAlmostEqual(np.sort(np.exp(self.x.value))[-2:] @ np.sort(w), 2)
+
+    def test_copy(self):
+        # Test copy
+        w = np.array([1, 2])
+        atom = cp.dotsort(self.x, w)
+        copy = atom.copy()
+        self.assertTrue(type(copy) is type(atom))
+        # A new object is constructed, so copy.args == atom.args but copy.args
+        # is not atom.args.
+        self.assertEqual(copy.args, atom.args)
+        self.assertFalse(copy.args is atom.args)
+        self.assertTrue(copy.args[0] is atom.args[0])
+        self.assertTrue(copy.args[1] is atom.args[1])
+
+        # Test copy with new args
+        copy = atom.copy(args=[self.x, w])
+        self.assertFalse(copy.args is atom.args)
+        self.assertTrue(copy.args[0] is atom.args[0])
+        self.assertFalse(copy.args[1] is atom.args[1])
+
+    def test_non_fixed_x(self):
+        r = np.array([2, 1, 0, -1, -1])
+        w = np.array([1.2, 1.1])
+        expr = cp.dotsort(self.x, w)
+        prob = cp.Problem(cp.Maximize(r @ self.x), [0 <= self.x, expr <= 1, cp.sum(self.x) == 1])
+        prob.solve()
+        self.assertAlmostEqual(expr.value, 1)
+        self.assertAlmostEqual(self.x.value[:2] @ w, 1)
+
+        # Test non-fixed x - unordered w
+        r = np.array([2, 1, 0, -1, -1])
+        w = np.array([1.2, 1.1, 1.3])
+        expr = cp.dotsort(self.x, w)
+        prob = cp.Problem(cp.Maximize(r @ self.x), [0 <= self.x, expr <= 1, cp.sum(self.x) == 1])
+        prob.solve()
+        self.assertAlmostEqual(expr.value, 1)
+        self.assertAlmostEqual(np.sort(self.x.value)[-3:] @ np.sort(w), 1)
+
+    def test_exceptions(self):
+        # len(w) > len(x)
+        with self.assertRaises(Exception) as cm:
+            cp.dotsort(self.x, [1, 2, 3, 4, 5, 8])
+        self.assertEqual(str(cm.exception),
+                         "The size of of W must be less or equal to the size of X.")
+
+        # two variable expressions
+        with self.assertRaises(Exception) as cm:
+            cp.dotsort(self.x, cp.Variable(3))
+        self.assertEqual(str(cm.exception),
+                         "The W argument must be constant.")
+
+        # swapped arguments
+        with self.assertRaises(Exception) as cm:
+            cp.dotsort([1, 2, 3], self.x)
+        self.assertEqual(str(cm.exception),
+                         "The W argument must be constant.")
+
+        # non-dcp composition
+        with self.assertRaises(Exception) as cm:
+            cp.Problem(cp.Minimize(cp.dotsort(cp.abs(self.x), [-1, 1]))).solve()
+        assert "Problem does not follow DCP rules" in str(cm.exception)
+
+        # # non-dpp composition
+        p = cp.Parameter(value=2)
+        p_squared = p ** 2
+        with self.assertRaises(Exception) as cm:
+            cp.Problem(cp.Minimize(cp.dotsort(self.x, p_squared))).solve(enforce_dpp=True)
+        assert "You are solving a parameterized problem that is not DPP" in str(cm.exception)
