@@ -74,6 +74,8 @@ atoms_minimize = [
     (cp.diag, (2, 2), [[-5, 1]], Constant([[-5, 0], [0, 1]])),
     (cp.exp, (2, 2), [[[1, 0], [2, -1]]],
      Constant([[math.e, 1], [math.e**2, 1.0 / math.e]])),
+    (lambda x: cp.xexp(cp.pos(x)), (2, 2), [[[1, 0], [2, .5]]],
+     Constant([[math.e, 0], [2 * math.e**2, 0.5 * math.e**.5]])),
     (cp.huber, (2, 2), [[[0.5, -1.5], [4, 0]]],
      Constant([[0.25, 2], [7, 0]])),
     (lambda x: cp.huber(x, 2.5), (2, 2), [[[0.5, -1.5], [4, 0]]],
@@ -88,11 +90,29 @@ atoms_minimize = [
     (cp.rel_entr, tuple(), [math.e, 1], Constant([math.e])),
     (cp.rel_entr, tuple(), [math.e, math.e], Constant([0])),
     (cp.rel_entr, (2,), [[math.e, 1], 1], Constant([math.e, 0])),
-    (lambda x: cp.kron(np.array([[1, 2], [3, 4]]), x), (4, 4), [np.array([[5, 6], [7, 8]])],
+    # kron with variable in the right operand
+    (lambda x: cp.kron(np.array([[1, 2], [3, 4]]), x), (4, 4),
+     [np.array([[5, 6], [7, 8]])],
      Constant(np.kron(np.array([[1, 2], [3, 4]]), np.array([[5, 6], [7, 8]])))),
+    (lambda x: cp.kron(np.array([[1, 2], [3, 4], [5, 6]]), x), (6, 4),
+     [np.array([[5, 6], [7, 8]])],
+     Constant(np.kron(np.array([[1, 2], [3, 4], [5, 6]]), np.array([[5, 6], [7, 8]])))),
+    (lambda x: cp.kron(np.array([[1, 2], [3, 4]]), x), (6, 4),
+     [np.array([[5, 6], [7, 8], [9, 10]])],
+     Constant(np.kron(np.array([[1, 2], [3, 4]]), np.array([[5, 6], [7, 8], [9, 10]])))),
+
+    # kron with variable in the left operand
+    (lambda x: cp.kron(x, np.array([[1, 2], [3, 4]])), (4, 4),
+     [np.array([[5, 6], [7, 8]])],
+     Constant(np.kron(np.array([[5, 6], [7, 8]]), np.array([[1, 2], [3, 4]])))),
+    (lambda x: cp.kron(x, np.array([[1, 2], [3, 4], [5, 6]])), (6, 4),
+     [np.array([[5, 6], [7, 8]])],
+     Constant(np.kron(np.array([[5, 6], [7, 8]]), np.array([[1, 2], [3, 4], [5, 6]])))),
+    (lambda x: cp.kron(x, np.array([[1, 2], [3, 4]])), (6, 4),
+     [np.array([[5, 6], [7, 8], [9, 10]])],
+     Constant(np.kron(np.array([[5, 6], [7, 8], [9, 10]]), np.array([[1, 2], [3, 4]])))),
     (cp.lambda_max, tuple(), [[[2, 0], [0, 1]]], Constant([2])),
     (cp.lambda_max, tuple(), [[[2, 0, 0], [0, 3, 0], [0, 0, 1]]], Constant([3])),
-
     (cp.lambda_max, tuple(), [[[5, 7], [7, -3]]], Constant([9.06225775])),
     (lambda x: cp.lambda_sum_largest(x, 2), tuple(),
      [[[1, 2, 3], [2, 4, 5], [3, 5, 6]]], Constant([11.51572947])),
@@ -144,6 +164,8 @@ atoms_minimize = [
     (lambda x: cp.norm(x, "nuc"), tuple(), [[[3, 4, 5], [6, 7, 8]]],
      Constant([14.618376738088918])),
     (lambda x: cp.sum_largest(cp.abs(x), 3), tuple(), [[1, 2, 3, -4, -5]], Constant([5 + 4 + 3])),
+    (lambda x: cp.dotsort(cp.abs(x), np.array([2, 1, 0.1])), tuple(), [[1, 2, 3, -4, -5]],
+     Constant([10 + 4 + .3])),
     (lambda x: cp.mixed_norm(x, 1, 1), tuple(), [[[1, 2], [3, 4], [5, 6]]],
      Constant([21])),
     (lambda x: cp.mixed_norm(x, 1, 1), tuple(), [[[1, 2, 3], [4, 5, 6]]],
