@@ -19,10 +19,11 @@ from __future__ import absolute_import, division, print_function
 import warnings
 
 import numpy as np
+import pytest
 import scipy.sparse as sp
 from numpy.testing import assert_allclose, assert_equal
 
-import cvxpy
+import cvxpy as cp
 from cvxpy.settings import EIGVAL_TOL
 from cvxpy.tests.base_test import BaseTest
 
@@ -56,16 +57,16 @@ class TestNonOptimal(BaseTest):
 
                     # Look for the extremum of the quadratic form
                     # under the simplex constraint.
-                    x = cvxpy.Variable(n)
+                    x = cp.Variable(n)
                     if action == 'minimize':
-                        q = cvxpy.quad_form(x, Q)
-                        objective = cvxpy.Minimize(q)
+                        q = cp.quad_form(x, Q)
+                        objective = cp.Minimize(q)
                     elif action == 'maximize':
-                        q = cvxpy.quad_form(x, -Q)
-                        objective = cvxpy.Maximize(q)
-                    constraints = [0 <= x, cvxpy.sum(x) == 1]
-                    p = cvxpy.Problem(objective, constraints)
-                    p.solve(solver=cvxpy.OSQP)
+                        q = cp.quad_form(x, -Q)
+                        objective = cp.Maximize(q)
+                    constraints = [0 <= x, cp.sum(x) == 1]
+                    p = cp.Problem(objective, constraints)
+                    p.solve(solver=cp.OSQP)
 
                     # check that cvxpy found the right answer
                     xopt = x.value.flatten()
@@ -77,46 +78,46 @@ class TestNonOptimal(BaseTest):
         """Test quad form with a sparse matrix.
         """
         Q = sp.eye(2)
-        x = cvxpy.Variable(2)
-        cost = cvxpy.quad_form(x, Q)
-        prob = cvxpy.Problem(cvxpy.Minimize(cost), [x == [1, 2]])
-        self.assertAlmostEqual(prob.solve(solver=cvxpy.OSQP), 5)
+        x = cp.Variable(2)
+        cost = cp.quad_form(x, Q)
+        prob = cp.Problem(cp.Minimize(cost), [x == [1, 2]])
+        self.assertAlmostEqual(prob.solve(solver=cp.OSQP), 5)
 
         # Here are our QP factors
-        A = cvxpy.Constant(sp.eye(4))
+        A = cp.Constant(sp.eye(4))
         c = np.ones(4).reshape((1, 4))
 
         # Here is our optimization variable
-        x = cvxpy.Variable(4)
+        x = cp.Variable(4)
 
         # And the QP problem setup
-        function = cvxpy.quad_form(x, A) - cvxpy.matmul(c, x)
-        objective = cvxpy.Minimize(function)
-        problem = cvxpy.Problem(objective)
+        function = cp.quad_form(x, A) - cp.matmul(c, x)
+        objective = cp.Minimize(function)
+        problem = cp.Problem(objective)
 
-        problem.solve(solver=cvxpy.OSQP)
+        problem.solve(solver=cp.OSQP)
         self.assertEqual(len(function.value), 1)
 
     def test_param_quad_form(self) -> None:
         """Test quad form with a parameter.
         """
-        P = cvxpy.Parameter((2, 2), PSD=True)
+        P = cp.Parameter((2, 2), PSD=True)
         Q = np.eye(2)
-        x = cvxpy.Variable(2)
-        cost = cvxpy.quad_form(x, P)
+        x = cp.Variable(2)
+        cost = cp.quad_form(x, P)
         P.value = Q
-        prob = cvxpy.Problem(cvxpy.Minimize(cost), [x == [1, 2]])
+        prob = cp.Problem(cp.Minimize(cost), [x == [1, 2]])
         with warnings.catch_warnings():
             warnings.simplefilter('ignore')
-            self.assertAlmostEqual(prob.solve(solver=cvxpy.SCS), 5)
+            self.assertAlmostEqual(prob.solve(solver=cp.SCS), 5)
 
     def test_non_symmetric(self) -> None:
         """Test when P is constant and not symmetric.
         """
         P = np.array([[2, 2], [3, 4]])
-        x = cvxpy.Variable(2)
+        x = cp.Variable(2)
         with self.assertRaises(Exception) as cm:
-            cvxpy.quad_form(x, P)
+            cp.quad_form(x, P)
         self.assertTrue("P must be symmetric/Hermitian."
                         in str(cm.exception))
 
@@ -124,14 +125,14 @@ class TestNonOptimal(BaseTest):
         """Test error when P is symmetric but not definite.
         """
         P = np.array([[1, 0], [0, -1]])
-        x = cvxpy.Variable(2)
+        x = cp.Variable(2)
         # Forming quad_form is okay
         with warnings.catch_warnings():
             warnings.simplefilter("ignore")
-            cost = cvxpy.quad_form(x, P)
-        prob = cvxpy.Problem(cvxpy.Minimize(cost), [x == [1, 2]])
+            cost = cp.quad_form(x, P)
+        prob = cp.Problem(cp.Minimize(cost), [x == [1, 2]])
         with self.assertRaises(Exception) as cm:
-            prob.solve(solver=cvxpy.SCS)
+            prob.solve(solver=cp.SCS)
         self.assertTrue("Problem does not follow DCP rules."
                         in str(cm.exception))
 
@@ -139,36 +140,36 @@ class TestNonOptimal(BaseTest):
         """Test that PSD check when eigenvalue is exactly -EIGVAL_TOL
         """
         P = np.array([[-0.999*EIGVAL_TOL, 0], [0, 10]])
-        x = cvxpy.Variable(2)
+        x = cp.Variable(2)
         # Forming quad_form is okay
         with warnings.catch_warnings():
             warnings.simplefilter("ignore")
-            cost = cvxpy.quad_form(x, P)
-            prob = cvxpy.Problem(cvxpy.Minimize(cost), [x == [1, 2]])
-            prob.solve(solver=cvxpy.SCS)
+            cost = cp.quad_form(x, P)
+            prob = cp.Problem(cp.Minimize(cost), [x == [1, 2]])
+            prob.solve(solver=cp.SCS)
 
     def test_nsd_exactly_tolerance(self) -> None:
         """Test that NSD check when eigenvalue is exactly EIGVAL_TOL
         """
         P = np.array([[0.999*EIGVAL_TOL, 0], [0, -10]])
-        x = cvxpy.Variable(2)
+        x = cp.Variable(2)
         # Forming quad_form is okay
         with warnings.catch_warnings():
             warnings.simplefilter("ignore")
-            cost = cvxpy.quad_form(x, P)
-            prob = cvxpy.Problem(cvxpy.Maximize(cost), [x == [1, 2]])
-            prob.solve(solver=cvxpy.SCS)
+            cost = cp.quad_form(x, P)
+            prob = cp.Problem(cp.Maximize(cost), [x == [1, 2]])
+            prob.solve(solver=cp.SCS)
 
     def test_obj_eval(self) -> None:
         """Test case where objective evaluation differs from result.
         """
-        x = cvxpy.Variable((2, 1))
+        x = cp.Variable((2, 1))
         A = np.array([[1.0]])
         B = np.array([[1.0, 1.0]]).T
         obj0 = -B.T @ x
-        obj1 = cvxpy.quad_form(B.T @ x, A)
-        prob = cvxpy.Problem(cvxpy.Minimize(obj0 + obj1))
-        prob.solve(solver=cvxpy.SCS)
+        obj1 = cp.quad_form(B.T @ x, A)
+        prob = cp.Problem(cp.Minimize(obj0 + obj1))
+        prob.solve(solver=cp.SCS)
         self.assertAlmostEqual(prob.value, prob.objective.value)
 
     def test_zero_term(self) -> None:
@@ -176,27 +177,45 @@ class TestNonOptimal(BaseTest):
         """
         data_norm = np.random.random(5)
         M = np.random.random(5*2).reshape((5, 2))
-        c = cvxpy.Variable(M.shape[1])
+        c = cp.Variable(M.shape[1])
         lopt = 0
         laplacian_matrix = np.ones((2, 2))
-        design_matrix = cvxpy.Constant(M)
-        objective = cvxpy.Minimize(
-            cvxpy.sum_squares(design_matrix @ c - data_norm) +
-            lopt * cvxpy.quad_form(c, laplacian_matrix)
+        design_matrix = cp.Constant(M)
+        objective = cp.Minimize(
+            cp.sum_squares(design_matrix @ c - data_norm) +
+            lopt * cp.quad_form(c, laplacian_matrix)
         )
         constraints = [(M[0] @ c) == 1]  # (K * c) >= -0.1]
-        prob = cvxpy.Problem(objective, constraints)
-        prob.solve(solver=cvxpy.SCS)
+        prob = cp.Problem(objective, constraints)
+        prob.solve(solver=cp.SCS)
 
     def test_zero_matrix(self) -> None:
         """Test quad_form with P = 0.
         """
-        x = cvxpy.Variable(3)
+        x = cp.Variable(3)
         A = np.eye(3)
         b = np.ones(3,)
         c = -np.ones(3,)
         P = np.zeros((3, 3))
-        expr = (1/2) * cvxpy.quad_form(x, P) + c.T @ x
-        prob = cvxpy.Problem(cvxpy.Minimize(expr),
-                             [A @ x <= b])
-        prob.solve(solver=cvxpy.SCS)
+        expr = (1/2) * cp.quad_form(x, P) + c.T @ x
+        prob = cp.Problem(cp.Minimize(expr),
+                          [A @ x <= b])
+        prob.solve(solver=cp.SCS)
+
+    def test_assume_psd(self) -> None:
+        """Test assume_PSD argument.
+        """
+        x = cp.Variable(3)
+        A = np.eye(3)
+        expr = cp.quad_form(x, A, assume_PSD=True)
+        assert expr.is_convex()
+
+        A = -np.eye(3)
+        expr = cp.quad_form(x, A, assume_PSD=True)
+        assert expr.is_convex()
+
+        prob = cp.Problem(cp.Minimize(expr))
+        # Transform to a SolverError.
+        with pytest.raises(cp.SolverError,
+                           match="Workspace allocation error!"):
+            prob.solve(solver=cp.OSQP)
