@@ -160,7 +160,7 @@ html_theme_options = {
 
 # The name for this set of Sphinx documents.  If None, it defaults to
 # "<project> v<release> documentation".
-#html_title = None
+html_title = f'CVXPY {version} documentation'
 
 # A shorter title for the navigation bar.  Default is the same as html_title.
 #html_short_title = None
